Extract AppDataDir from LevelDbPersistence
