+++ Firestore/core/src/firebase/firestore/local/CMakeLists.txt
@@ -50,12 +50,10 @@ cc_library(
     leveldb_migrations.cc
     leveldb_migrations.h
     leveldb_mutation_queue.cc
-    leveldb_persistence.cc
     leveldb_mutation_queue.h
+    leveldb_persistence.cc
     leveldb_persistence.h
     leveldb_persistence_apple.mm
-    leveldb_persistence_posix.cc
-    leveldb_persistence_win.cc
     leveldb_query_cache.cc
     leveldb_query_cache.h
     leveldb_remote_document_cache.cc
