+++ Firestore/core/src/firebase/firestore/local/leveldb_persistence_apple.mm
@@ -29,28 +29,6 @@ namespace local {
 
 using util::Path;
 using util::Status;
-using util::StatusOr;
-
-StatusOr<Path> LevelDbPersistence::AppDataDirectory() {
-#if TARGET_OS_IOS
-  NSArray<NSString*>* directories = NSSearchPathForDirectoriesInDomains(
-      NSDocumentDirectory, NSUserDomainMask, YES);
-  return Path::FromNSString(directories[0]).AppendUtf8(kReservedPathComponent);
-
-#elif TARGET_OS_TV
-  NSArray<NSString*>* directories = NSSearchPathForDirectoriesInDomains(
-      NSCachesDirectory, NSUserDomainMask, YES);
-  return Path::FromNSString(directories[0]).AppendUtf8(kReservedPathComponent);
-
-#elif TARGET_OS_OSX
-  std::string dot_prefixed = absl::StrCat(".", kReservedPathComponent);
-  return Path::FromNSString(NSHomeDirectory()).AppendUtf8(dot_prefixed);
-
-#else
-#error "Don't know where to store documents on this platform."
-
-#endif
-}
 
 Status LevelDbPersistence::ExcludeFromBackups(const Path& dir) {
   NSURL* dir_url = [NSURL fileURLWithPath:dir.ToNSString()];
