+++ Firestore/core/src/firebase/firestore/util/filesystem_win.cc
@@ -19,6 +19,7 @@
 #if defined(_WIN32)
 
 #include <windows.h>
+#include <Shlobj.h>
 
 #include <cerrno>
 #include <string>
@@ -43,6 +44,21 @@ Status IsDirectory(const Path& path) {
   }
 
   return Status{Error::FailedPrecondition, path.ToUtf8String()};
+}
+
+StatusOr<Path> AppDataDir(absl::string_view app_name) {
+  wchar_t* path = nullptr;
+  HRESULT hr = SHGetKnownFolderPath(FOLDERID_LocalAppData, 0, nullptr, &path);
+  if (FAILED(hr)) {
+    CoTaskMemFree(path);
+    return Status::FromLastError(
+        HRESULT_CODE(hr),
+        "Failed to find the local application data directory");
+  }
+
+  Path result = Path::FromUtf16(path, wcslen(path)).AppendUtf8(app_name);
+  CoTaskMemFree(path);
+  return std::move(result);
 }
 
 Path TempDir() {
